Removed all old extractors and used new extractor
